refactor(object_data): extract TLS types into thread_local;
